+++ src/Site/Urls/ChiefResponse.php
@@ -4,12 +4,25 @@ declare(strict_types=1);
 
 namespace Thinktomorrow\Chief\Site\Urls;
 
+use Throwable;
+use Illuminate\Support\Arr;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Str;
+use Illuminate\Auth\AuthenticationException;
+use Illuminate\Session\TokenMismatchException;
+use Illuminate\Validation\ValidationException;
+use Illuminate\Database\RecordsNotFoundException;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Http\Exceptions\HttpResponseException;
+use Illuminate\Database\MultipleRecordsFoundException;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Symfony\Component\HttpKernel\Exception\HttpException;
 use Symfony\Component\HttpFoundation\Response as BaseResponse;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Thinktomorrow\Chief\Shared\ModelReferences\ModelReference;
 use Thinktomorrow\Chief\Site\Visitable\Visitable;
+use Illuminate\Routing\Exceptions\BackedEnumCaseNotFoundException;
+use Symfony\Component\HttpFoundation\Exception\SuspiciousOperationException;
 
 final class ChiefResponse
 {
@@ -32,12 +45,29 @@ final class ChiefResponse
 
             return static::findModel($urlRecord)->response();
         } catch (\Throwable $e) {
-            if (config('chief.strict')) {
+            if (config('chief.strict') || !static::shouldBeIgnored($e)) {
                 throw $e;
             }
         }
 
         throw new NotFoundHttpException('No url or model found for request [' . $slug . '] for locale [' . $locale . '].');
+    }
+
+    private static function shouldBeIgnored(Throwable $e): bool
+    {
+        return ! is_null(Arr::first(static::ignoredExceptions(), fn ($type) => $e instanceof $type));
+    }
+
+    private static function ignoredExceptions(): array
+    {
+        return [
+            UrlRecordNotFound::class,
+            AuthenticationException::class,
+            AuthorizationException::class,
+            HttpException::class,
+            HttpResponseException::class,
+            ModelNotFoundException::class,
+        ];
     }
 
     private static function findModel(UrlRecord $urlRecord): Visitable